Add Typesense search refinements
