+++ .nuxtrc
