+++ app/components/HomeHero.vue
@@ -9,14 +9,12 @@ function openSearch() {
 	<section
 		class="not-prose relative mt-6 mb-8 overflow-hidden rounded-2xl bg-neutral-950 text-white shadow-[0_10px_24px_-16px_rgba(0,0,0,0.4)] ring-1 ring-white/[0.06]"
 	>
-		<!-- Painting layer fills container, fades from top to bottom -->
 		<div class="hero-mask absolute inset-0">
 			<ClientOnly>
 				<HomeHeroShader :anchor-y="0.5" :wash-amount="0" />
 			</ClientOnly>
 		</div>
 
-		<!-- Debug panel target (outside mask so it's not faded out) -->
 		<div
 			id="hero-shader-debug-slot"
 			class="absolute top-0 right-0 z-30"
@@ -34,17 +32,33 @@ function openSearch() {
 			</p>
 			<button
 				type="button"
-				class="mt-7 flex w-full max-w-md cursor-pointer items-center gap-3 rounded-full border border-white/[0.08] bg-white/[0.04] px-4 py-3 text-left text-sm text-white/60 transition duration-150 hover:border-primary/45 hover:bg-white/[0.06] hover:ring-4 hover:ring-primary/10"
+				class="home-hero-search mt-7 flex w-full max-w-md cursor-pointer items-center gap-3 rounded-full px-4 py-3 text-left text-sm text-white/60"
 				aria-label="Search documentation"
 				@click="openSearch"
 			>
-				<UIcon name="material-symbols:search" class="size-4 shrink-0 text-white/60" />
+				<UIcon
+					name="heroicons:magnifying-glass-20-solid"
+					class="size-4 shrink-0 text-white/60"
+				/>
 				<span class="flex-1">Search the docs...</span>
 				<kbd class="rounded border border-white/10 bg-white/10 px-1.5 py-0.5 font-mono text-xs text-white/70">⌘K</kbd>
 			</button>
 			<div class="mt-6 flex flex-wrap gap-2.5">
-				<UButton label="Get Started" trailing-icon="material-symbols:arrow-forward" color="primary" size="lg" to="/getting-started/overview" />
-				<UButton label="What is Directus?" trailing-icon="material-symbols:play-arrow-outline" color="neutral" variant="subtle" size="lg" to="/getting-started/overview" />
+				<UButton
+					label="Get Started"
+					trailing-icon="i-ph-arrow-right"
+					color="primary"
+					size="lg"
+					to="/getting-started/overview"
+				/>
+				<UButton
+					label="What is Directus?"
+					trailing-icon="i-ph-play"
+					color="neutral"
+					variant="subtle"
+					size="lg"
+					to="/getting-started/overview"
+				/>
 			</div>
 		</div>
 	</section>
@@ -66,5 +80,17 @@ function openSearch() {
 	mask-repeat: no-repeat;
 	-webkit-mask-size: 100% 100%;
 	mask-size: 100% 100%;
+}
+
+.home-hero-search {
+	background: rgba(255, 255, 255, 0.04);
+	border: 1px solid rgba(255, 255, 255, 0.08);
+	transition: border-color 0.15s ease, background-color 0.15s ease, box-shadow 0.15s ease;
+}
+
+.home-hero-search:hover {
+	background: rgba(255, 255, 255, 0.06);
+	border-color: color-mix(in oklab, var(--color-primary) 45%, transparent);
+	box-shadow: 0 0 0 4px color-mix(in oklab, var(--color-primary) 12%, transparent);
 }
 </style>
