+++ app/composables/useSearchOverlay.ts
@@ -1,10 +1,7 @@
 import { createSharedComposable } from '@vueuse/core';
+import { LazyDocsSearchPalette } from '#components';
 
 export const useSearchOverlay = createSharedComposable(() => {
-	function open() {
-		if (!import.meta.client) return;
-		document.querySelector<HTMLButtonElement>('#docsearch .DocSearch-Button')?.click();
-	}
-
-	return { open };
+	const overlay = useOverlay();
+	return overlay.create(LazyDocsSearchPalette);
 });
